Delete account/authenticate.py file
